Allow control of PHP type declaration generation
This can be controlled with %feature("php:type"), which being a SWIG
feature can be specified globally, per class, per method, etc.

If unset or set to "0" then no type declarations are generated.

If set to "1" then type declarations are generated for both
parameters and return types.

Setting it to "compat" is the same as "1" except no return type
declarations are generated for virtual methods for which directors
are enabled.  This provides better compatibility for PHP subclasses
of wrapped virtual methods in existing SWIG-generated bindings