Fix invalid code in testcase typemap_qualifier_strip.i
This testcase was using a pointer to a variable which had gone out
of scope, which is undefined behaviour.  The test was only passing
because the tests are compiled without optimisation by default and
the memory where this value lived happened to remain intact in that
case with current versions of compilers we test with.  Running the
testsuite with CXXFLAGS=-O2 causes this test to fail.

Fix by extending the lifetime of the variable to the whole wrapper
function.

Fixes #1925