Fix friends_runme.*
Most of these test D_d twice when they really should be testing
D_d once and D_i once (the variable name is `di` and the values
assigned are integers).

This was wrong in the initial version for Python in
708021a809a566a410ec0e6bf5c63029392315e1 and it looks like subsequent
additions for other languages have just copied that mistake.