Reenable and fix octave horzcat operator test
This test was disabled with 0a0743f25cc0bbb395d03a27ac62887c8cbee5cf since
it fails with octave 7.2.0

The test function horzcat now uses a variable length argument list to fix this.
Before octave 7 it seems to be possible to call a N-argument octave function
with > N arguments without any error. With octave 7 this seems no longer to be
possible which caused the test failure.