[ci] Fix CI to actually test PHP 8.0; add PHP 8.2
The CI job for PHP 8.0 was relying on not specifying a version
giving us PHP 8.0, but actually it gives us 8.1 currently.

This seems too brittle, so always specify the version explicitly as
at worst it means uninstalling and reinstalling PHP packages in a
case we could avoid.

Also add PHP 8.2 testing.

See #2383