python/cast_mode: Fix integer overflow of LONG_MAX+1.
In SWIG_CanCastAsInteger, we check for double(value) <= double(LONG_MAX).

However LONG_MAX cannot be represented as double if sizeof(long)=8, so it is usually rounded
to LONG_MAX+1. So SWIG_CanCastAsInteger returned true for LONG_MAX+1 and the next cast ended
up in an integer overflow.

Worse for unsigned long: It is called by smaller unsigned integers with range checks after
the convertion to unsigned long. As the integer overflow yields a 0x0, this is valid for all
small unsigned integers. Smaller signed integers will always fail this range check after the
integer overflow.

This fixes the tests on my machine:
  as_l(lmaxd + LSB)
  as_ll(llmaxd + LSB)
  as_ul(ulmax + LSB)
  as_ull(ullmax + LSB)

Wrong old results before this commit:
  as_l(float(2**63)) := -0x8000000000000000
  as_ll(float(2**63)) := -0x8000000000000000
  as_uc(float(2**64)) := 0x0
  as_us(float(2**64)) := 0x0
  as_ui(float(2**64)) := 0x0
  as_ul(float(2**64)) := 0x0
  as_ull(float(2**64)) := 0x0

Now all of them throw an exception.