Improved error checking when defining classes and using %template.
1. When a template is instantiated via %template and uses the unary scope
operator ::, an error occurs if the instantiation is attempted within a
namespace that does not enclose the instantiated template.
For example, the following will now error as ::test::max is not enclosed within test1:

Error: '::test::max' resolves to 'test::max' and was incorrectly instantiated in
scope 'test1' instead of within scope 'test'.
  namespace test1 {
    %template(maxchar) ::test::max<char>;
  }

2. SWIG previously failed to always detect a template did not exist when using
%template. In particular when instantiating a global template incorrectly within
namespace. The code below now correctly emits an error:

Error: Template 'test5::GlobalVector' undefined.
  namespace test5 {
  }
  template<typename T> struct GlobalVector {};
  %template(GVI) test5::GlobalVector<int>;

3. Also error out if an attempt is made to define a class using the unary scope
operator ::. The following is not legal C++ and now results in an error:

Error: Using the unary scope operator :: in class definition '::Space2::B' is invalid.
  namespace Space2 {
    struct B;
  }
  struct ::Space2::B {};