[ci] Disable ubuntu-18.04 jobs
Github has now dropped support for running jobs on Ubuntu 18.04 and
these jobs now sit un-run for a day and then fail.

Hopefully many of these jobs can be resurrected on a newer Ubuntu
version, but as a first step let's get git master back to a state where
all the enabled jobs can be expected to pass.