Python: Suggest argcargv.i to handle char**
We were providing an example set of typemaps in the manual, but they
were specific to Python2 which isn't helpful these days.

For typical cases argcargv.i is a better option.  It doesn't currently
seem to directly support the "argv without argc" case which this example
actually shows, but generally APIs take a length as well as a char**.

Closes: #2040