Make typemaps consistently use string::data() vs c_str()
Use c_str() only when we need a terminating zero byte and data() when
we don't (as we already did in most cases).  The two methods in fact
do the same thing as of C++11 (and in practice did for all C++98
implementations I'm aware of) but it's useful to make clear when we need
a terminating zero byte and when we don't, for example for adding
string_view support (string_view supports data(), but can't support
c_str() (since it could be a slice from the middle of a std::string).