Don't test string_view global or member vars
These can't really work reliably because the C/C++ variable stores
a pointer to string data in the scripting language.  For Lua and
PHP that happens to work (unless some sort of GC compaction pass
moves the string) but for Java the passed string data is a temporary
UTF8(ish) conversion of the native string format.

We can still usefully test const global and member variables.