Skip declarations SWIG can't parse
There are still many valid C++ constructs which SWIG can't parse
(typically from newer C++ standards) which currently halt SWIG with a
parse error.

We've had quite a lot of user reports where this happens but the
user doesn't actually want to wrap the problematic declaration, so
let's try getting SWIG to emit a warning and skip to the next
declaration.