Fix bogus code in C++11 trailing return type handling
The parser rule for this has clearly been adapted from the rule for
traditional function syntax.  The "val" and "bitfield" attributes
that the traditional rule sets on the resulting node had been
changed here to get copied off the cpp_const token, which doesn't
explicitly set either, so we pick up whatever values were left over
in that memory from previous use of that entry in the parser's
stack.

This could potentially lead to incorrect behaviour, but I spotted
this while reading the code and I don't have an example that actually
goes wrong with this.

Related to #2612