Require Bison 3.5
We now require Bison (previously we worked with some other yacc
implementations, though they did need to support the non-POSIX
%expect feature), and require at version 3.0 or higher (ideally
we'd require 3.5 or higher, but we need to sort out the appveyor
config before we can do that).  This will allow us to make use of some
Bison-specific features which will make the parser easier to extend and
maintain.

See #2612