Make use of Bison's default action being $$ = $1;
This is conceptually very natural and makes the parser grammar a
little easier to read.

It also results in a smaller generated parser (it seems Bison knows it
can just leave the value on the top of the stack for the default action,
but lacks an optimisation to know it can do this for an explicit action
of `$$ = $1;`).

See #2612