Stop defining yylex to itself
There are no checks for yylex being defined in the generated
parser so this seems to serve no purpose.  It's been present
since before the start of version control history so presumably
was needed with some ancient bison or yacc version, or else
morphed out of `#define yylex some_other_name`.