Eliminate duplicate scanner codes
Having more than one name for the same token is unnecessary and makes
the code harder to follow.

These aliases were also mostly unused:

SWIG_TOKEN_TIMES was defined to the same value as SWIG_TOKEN_STAR,
but is not used anywhere.

SWIG_TOKEN_DIVIDE was defined to the same value as SWIG_TOKEN_SLASH but
also not used.

SWIG_TOKEN_MODULO was defined to the same value as SWIG_TOKEN_PERCENT
but only used in one place so just use SWIG_TOKEN_PERCENT there too.