Fail at compile time for multiple typemap substitution
Define a goto label in the typemap with a fixed name which will give a
multiple definition error if the typemap is substituted more than once
in a particular wrapper function.

This way we'll also catch this potential bug for target languages that
don't have a runme for the ignore_parameter testcase.