Missing scope restore in Allocate::classDeclaration
Restoring the scope was missing when handling base classes - the scope
was not set correctly after visiting base classes.

This bug doesn't seem to manifest itself in any way as either the code
uses fully resolved types or specifically sets the scope when needed,
so I can't see anything that is fixed by this.